Fixed Buffer Cast Typeing Error

+++ kernel/kernel/src/fileSystems/fat32/fat32.cpp
@@ -15,10 +15,10 @@ void memset(void* start, uint8_t value, uint64_t num)
 void read_sector(unsigned int lba, int num, int selector, void *dst) {
     DiskInterface::SataDiskInterface tmpsdi(osData.tmp_port);
     // 计算起始扇区号
-    uint64_t sector = static_cast<uint64_t>(lba) + selector;
+    uint64_t sector = ((uint64_t)lba + (uint64_t)selector);
     
     // 计算读取的扇区数量
-    uint32_t sectorCount = static_cast<uint32_t>(num);
+    uint32_t sectorCount = (uint32_t)(num);
     
     // 调用SATARead函数来实现读取扇区的操作
     // 这里假设SATARead函数实现良好且返回一个bool值表示操作成功与否
